% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell.R
\name{assign_reactions_to_flow_cells_reaction_centric}
\alias{assign_reactions_to_flow_cells_reaction_centric}
\title{Expects a tibble of libraries to seq with following columns: ATAC/GEX_extra_mil reads, TENX_reaction_ID, ATAC/GEX_index}
\usage{
assign_reactions_to_flow_cells_reaction_centric(
  type,
  flow_cell_count,
  mil_read_limits,
  libraries_to_seq_tibble_in,
  names,
  n_attempts = 50,
  verbose = TRUE
)
}
\arguments{
\item{type}{ATAC or GEX}

\item{flow_cell_count}{numeric}

\item{mil_read_limits}{vector of read capacity of flow cells in millions corresponding with length equal to flow_cell_count}

\item{libraries_to_seq_tibble_in}{tibble of libraries to seq with following columns: ATAC/GEX_extra_mil reads, TENX_reaction_ID, ATAC/GEX_index}

\item{names}{flow cell names/identifiers}

\item{n_attempts}{how many attempts to solve the puzzle}

\item{verbose}{verbosity TRUE/FALSE}
}
\value{
Combined tible with flow cell identifier
}
\description{
Expects a tibble of libraries to seq with following columns: ATAC/GEX_extra_mil reads, TENX_reaction_ID, ATAC/GEX_index
}
