% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOP_mRNA_seq.R
\name{extract_specific_intersection}
\alias{extract_specific_intersection}
\title{Get specific intersection}
\usage{
extract_specific_intersection(named_list, extract_vectors)
}
\arguments{
\item{named_list}{A named list of entities}

\item{extract_vector}{A list of extrcation vectors, i.e. a list of named boolean vectors,
where the names correspond to the given set, and the boolean whether it should be in that set or not}
}
\description{
Get specific intersection
}
