% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOP_mRNA_seq.R
\name{generate_gene_info_list}
\alias{generate_gene_info_list}
\title{Generate external gene names for ENSEMBL-IDS.}
\usage{
generate_gene_info_list(dataset = "hsapiens_gene_ensembl", gene_counts_raw)
}
\arguments{
\item{dataset}{Choose human or mouse biomart ensembl dataset}

\item{gene_counts_raw}{Raw salmon output with ensemble IDS as rownames.}
}
\value{
nothing
}
\description{
Generate external gene names for ENSEMBL-IDS.
}
