% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOP_mRNA_seq.R
\name{get_DE_datatable}
\alias{get_DE_datatable}
\title{Return a nicely formatted DT::datatable object of the DE results}
\usage{
get_DE_datatable(df, ...)
}
\arguments{
\item{df}{The dataframe with the DE results}

\item{...}{extra arguments passed to DT::datatable()}
}
\value{
A DT::datatable
}
\description{
Return a nicely formatted DT::datatable object of the DE results
}
