% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOP_mRNA_seq.R
\name{get_intersect_members}
\alias{get_intersect_members}
\title{Function to get the intersection values correpsodning to a UpsetR-plot}
\usage{
get_intersect_members(x, ...)
}
\arguments{
\item{...}{The unique combination of intersections to get the result for}

\item{The}{binary membership df, can be geneated from named list via fromList()}
}
\description{
Function to get the intersection values correpsodning to a UpsetR-plot
}
