% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOP_mRNA_seq.R
\name{get_sig_entities_between_coefficients_across_tissues}
\alias{get_sig_entities_between_coefficients_across_tissues}
\title{Gets a nested list of significant entities (genes or GO-terms) between coefficients across tissues.
Assumes there is a data dir with each tissue described in its name, with GO and DE-tables in them
with corresponding names.}
\usage{
get_sig_entities_between_coefficients_across_tissues(
  data_dir,
  type,
  tissues,
  coefficients,
  genes_type,
  verbose = FALSE
)
}
\arguments{
\item{type}{Either "DE" or "GO" depending on the type of entity}

\item{tissues}{Character vector: Comma-seperated list of tissues as present in the project root, i.e. Hyp,NAc}
}
\value{
A nested list, where the upper levels corresponds to supplied  coefficients,
and the lower levels each contain a vector of significant gene names or GO-terms for the tissue
}
\description{
Gets a nested list of significant entities (genes or GO-terms) between coefficients across tissues.
Assumes there is a data dir with each tissue described in its name, with GO and DE-tables in them
with corresponding names.
}
