% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOP_mRNA_seq.R
\name{get_sig_entities_between_tissues_across_coefficients}
\alias{get_sig_entities_between_tissues_across_coefficients}
\title{Gets a nested list of significant entities (genes or GO-terms) between tissues across coefficients.
Assumes there is a data dir with each tissue described in its name, with GO and DE-tables in them
with corresponding names.}
\usage{
get_sig_entities_between_tissues_across_coefficients(
  data_dir,
  type,
  coefficients,
  tissues,
  genes_type
)
}
\arguments{
\item{type}{Either "DE" or "GO" depending on the type of entity}

\item{coefficients}{Character vector: Comma-separated list of coefficients as present in the data directory, i.e. Group2,Group3}

\item{tissues}{Character vector: Comma-seperated list of tissues as present in the project root, i.e. Hyp,NAc}

\item{genes_type}{Either ENSEMBL or NCBI}
}
\value{
A nested list, where the upper levels corresponds to supplied tissues,
and the lower levels each contain a vector of significant gene names or GO-terms for the given condition
}
\description{
Gets a nested list of significant entities (genes or GO-terms) between tissues across coefficients.
Assumes there is a data dir with each tissue described in its name, with GO and DE-tables in them
with corresponding names.
}
