% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOP_mRNA_seq.R
\name{get_translate_vec}
\alias{get_translate_vec}
\title{Get a named vector that can be used to translate ENSEMBL-ids into gene synmbols}
\usage{
get_translate_vec(dataset = "hsapiens_gene_ensembl", gene_counts_raw_path)
}
\arguments{
\item{dataset}{The biomart dataset}

\item{gene_counts_raw_path}{the gne counts raw object to extract ENSEMBL_IDS from}
}
\value{
A named vector that can be used for translation
}
\description{
Get a named vector that can be used to translate ENSEMBL-ids into gene synmbols
}
