% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML_reports.R
\name{initialize_plotly_and_DT}
\alias{initialize_plotly_and_DT}
\title{Required initialization and loading Plotly and DT libaries}
\usage{
initialize_plotly_and_DT()
}
\value{
Nothing
}
\description{
Required initialization and loading Plotly and DT libaries
}
