% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOP_mRNA_seq.R
\name{make_venn}
\alias{make_venn}
\title{Draws a Venn diagram of significant entities (genes or GO-terms) between tissues across coefficients,
or between coefficients across tissues}
\usage{
make_venn(
  between_coefficients,
  type,
  tissues,
  coefficients,
  genes_type,
  data_dir
)
}
\arguments{
\item{between_coefficients}{boolean: Whether to plot overlaps between coefficients (TRUE) or between tissues (FALSE)}

\item{type}{character vector: Either "DE" or "GO" depending on the type of entity}

\item{tissues}{character vector: Comma-seperated list of tissues as present in the project root, i.e. Hyp, NAc}

\item{coefficients}{character vector: Comma-separated list of coefficients as present in the data directory, i.e. Group2,Group3}
}
\value{
A Venn Diagram of the ggplot-kind.
}
\description{
Draws a Venn diagram of significant entities (genes or GO-terms) between tissues across coefficients,
or between coefficients across tissues
}
