% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOP_mRNA_seq.R
\name{overlapGroups}
\alias{overlapGroups}
\title{Get intersection values from a named list, corresponding to the Upset-plot}
\usage{
overlapGroups(listInput, sort = TRUE)
}
\arguments{
\item{listInput}{Named list which can also be used in Upset fromList() function}

\item{sort}{whether to sort the output,}
}
\value{
Returns the labels of the intersecting values
}
\description{
Get intersection values from a named list, corresponding to the Upset-plot
}
