% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_processing.R
\name{switch_names_and_values_for_vec}
\alias{switch_names_and_values_for_vec}
\title{Switch names and values for named vector}
\usage{
switch_names_and_values_for_vec(named_vector)
}
\arguments{
\item{named_vector}{The named vector to switch names and values for}
}
\value{
Named vector with switched names and values
}
\description{
Switch names and values for named vector
}
