% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_processing.R
\name{write_tsv_and_excel}
\alias{write_tsv_and_excel}
\title{Write TSV and excel file}
\usage{
write_tsv_and_excel(x, path, ...)
}
\arguments{
\item{x}{Object to write to a TSV and excel file}

\item{path}{Path of the resulting files}
}
\value{
Nothing.
}
\description{
Write TSV and excel file
}
